Moved to pages the components that are now effectively pages
