+++ src/pages/trader/index.js
@@ -15,7 +15,7 @@ import {
 import SmallItemTable from '../../components/small-item-table/index.js';
 import QuestTable from '../../components/quest-table/index.js';
 import TraderResetTime from '../../components/trader-reset-time/index.js';
-import ErrorPage from '../../components/error-page/index.js';
+import ErrorPage from '../error-page/index.js';
 import LoadingSmall from '../../components/loading-small/index.js';
 import PropertyList from '../../components/property-list/index.js';
 import formatPrice from '../../modules/format-price.js';
